update and correct all snippets
