--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/Entity.cs
@@ -1,41 +1,0 @@
-﻿namespace ClassesAndObjects
-{
-    // <UsingEntity>
-    using System;
-    class EntityExample
-    {
-        public static void Usage()
-        {
-            Entity.SetNextSerialNo(1000);
-            Entity e1 = new Entity();
-            Entity e2 = new Entity();
-            Console.WriteLine(e1.GetSerialNo());            // Outputs "1000"
-            Console.WriteLine(e2.GetSerialNo());            // Outputs "1001"
-            Console.WriteLine(Entity.GetNextSerialNo());    // Outputs "1002"
-        }
-    }
-    // </UsingEntity>
-    // <EntityClass>
-    class Entity
-    {
-        static int nextSerialNo;
-        int serialNo;
-        public Entity()
-        {
-            serialNo = nextSerialNo++;
-        }
-        public int GetSerialNo()
-        {
-            return serialNo;
-        }
-        public static int GetNextSerialNo()
-        {
-            return nextSerialNo;
-        }
-        public static void SetNextSerialNo(int value)
-        {
-            nextSerialNo = value;
-        }
-    }
-    // </EntityClass>
-}
