--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/OutExample.cs
@@ -1,19 +1,0 @@
-﻿namespace ClassesAndObjects
-{
-    using System;
-    // <OutExample>
-    class OutExample
-    {
-        static void Divide(int x, int y, out int result, out int remainder)
-        {
-            result = x / y;
-            remainder = x % y;
-        }
-        public static void OutUsage()
-        {
-            Divide(10, 3, out int res, out int rem);
-            Console.WriteLine("{0} {1}", res, rem);	// Outputs "3 1"
-        }
-    }
-    // </OutExample>
-}
