--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/Pair.cs
@@ -1,10 +1,0 @@
-﻿namespace ClassesAndObjects
-{
-    // <DefinePairClass>
-    public class Pair<TFirst,TSecond>
-    {
-        public TFirst First;
-        public TSecond Second;
-    }
-    // </DefinePairClass>
-}
