--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/RefExample.cs
@@ -1,21 +1,0 @@
-﻿namespace ClassesAndObjects
-{
-    using System;
-    // <RefExample>
-    class RefExample
-    {
-        static void Swap(ref int x, ref int y)
-        {
-            int temp = x;
-            x = y;
-            y = temp;
-        }
-        public static void SwapExample()
-        {
-            int i = 1, j = 2;
-            Swap(ref i, ref j);
-            Console.WriteLine($"{i} {j}");    // Outputs "2 1"
-        }
-    }
-    // </RefExample>
-}
