--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/Squares.cs
@@ -1,20 +1,0 @@
-﻿namespace ClassesAndObjects
-{
-    using System;
-    // <SquaresClass>
-    class Squares
-    {
-        public static void WriteSquares()
-        {
-            int i = 0;
-            int j;
-            while (i < 10)
-            {
-                j = i * i;
-                Console.WriteLine($"{i} x {i} = {j}");
-                i = i + 1;
-            }
-        }
-    }
-    // </SquaresClass>
-}
