--- docs/csharp/tour-of-csharp/Snippets/classes-and-objects/access-by-line.txt
@@ -1,12 +1,0 @@
-Program.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Pair.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Point.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Color.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-RefExample.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-OutExample.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Squares.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Entity.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-Expressions.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md 
-Overloading.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-ListBasedExamples.cs: ~/docs/csharp/tour-of-csharp/classes-and-objects.md
-
