--- docs/csharp/tour-of-csharp/Snippets/interfaces/access-by-line.txt
@@ -1,1 +1,0 @@
-Program.cs: ~/docs/csharp/tour-of-csharp/interfaces.md
