+++ docs/csharp/tour-of-csharp/snippets/shared/AcmeStack.cs
@@ -1,5 +1,4 @@
-﻿// <SnippetStackExample>
-using System;
+﻿using System;
 namespace Acme.Collections
 {
     public class Stack<T>
@@ -32,26 +31,4 @@ namespace Acme.Collections
             }
         }
     }
-}
-// </SnippetStackExample>
-
-namespace StackExample
-{
-    // <SnippetStackUsage>
-    using System;
-    using Acme.Collections;
-    class Example
-    {
-        static void Main()
-        {
-            Stack<int> s = new Stack<int>();
-            s.Push(1); // stack contains 1
-            s.Push(10); // stack contains 1, 10
-            s.Push(100); // stack contains 1, 10, 100
-            Console.WriteLine(s.Pop()); // stack contains 1, 10
-            Console.WriteLine(s.Pop()); // stack contains 1
-            Console.WriteLine(s.Pop()); // stack is empty
-        }
-    }
-    // </SnippetStackUsage>
 }
