+++ docs/csharp/tour-of-csharp/snippets/shared/Program.cs
@@ -1,0 +1,14 @@
+﻿namespace TourOfCsharp
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            global::Example.Main();
+
+            Types.Examples();
+            ClassesObjects.Examples();
+            Features.Examples();
+        }
+    }
+}
