+++ docs/csharp/tour-of-csharp/snippets/shared/StackUsage.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Acme.Collections;
+class Example
+{
+    public static void Main()
+    {
+        Stack<int> s = new Stack<int>();
+        s.Push(1); // stack contains 1
+        s.Push(10); // stack contains 1, 10
+        s.Push(100); // stack contains 1, 10, 100
+        Console.WriteLine(s.Pop()); // stack contains 1, 10
+        Console.WriteLine(s.Pop()); // stack contains 1
+        Console.WriteLine(s.Pop()); // stack is empty
+    }
+}
