+++ docs/csharp/tour-of-csharp/snippets/shared/TourOfCsharp.csproj
@@ -1,8 +1,9 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <StartupObject>TourOfCsharp.Program</StartupObject>
   </PropertyGroup>
 
 </Project>
